code/requirements.txt

jsonschema[format-nongpl]==4.21.1 

influxdb-client[async,extra]==1.49.0 
pandas==2.3.0 
scipy==1.15.3 

paho-mqtt==2.1.0
aiohttp==3.12.12 
apscheduler==3.10.4
SQLAlchemy==2.0.41 